paperenvbench/tasks/videomae_classification_minimal/requirements_lock.txt
filelock==3.29.0
fsspec==2026.4.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==26.2
pillow==12.2.0
PyYAML==6.0.3
setuptools==82.0.1
sympy==1.14.0
timm==0.4.12
torch==2.4.1
torchvision==0.19.1
triton==3.0.0
typing_extensions==4.15.0
wheel==0.47.0
